move more types to primitive.py
