+++ spy/vm/list.py
@@ -1,8 +1,8 @@
 from typing import (TYPE_CHECKING, Any, no_type_check, Optional, Type, ClassVar,
                     TypeVar, Generic)
 from spy.fqn import QN
-from spy.vm.primitive import W_Void
-from spy.vm.object import (W_Object, spytype, W_Type, W_Dynamic, W_I32, W_Bool)
+from spy.vm.primitive import W_I32, W_Bool, W_Void
+from spy.vm.object import (W_Object, spytype, W_Type, W_Dynamic)
 from spy.vm.builtin import builtin_func
 if TYPE_CHECKING:
     from spy.vm.vm import SPyVM
