+++ spy/vm/modules/jsffi.py
@@ -1,10 +1,10 @@
 from typing import TYPE_CHECKING
 import struct
 from spy.fqn import QN
-from spy.vm.primitive import W_Void
+from spy.vm.primitive import W_F64, W_I32, W_Void
 from spy.vm.b import B
 from spy.vm.object import spytype, Member, Annotated
-from spy.vm.w import (W_Func, W_Type, W_Object, W_I32, W_F64, W_Str,
+from spy.vm.w import (W_Func, W_Type, W_Object, W_Str,
                       W_Dynamic, W_List, W_FuncType)
 from spy.vm.list import W_List
 from spy.vm.opimpl import W_OpImpl, W_OpArg
