+++ spy/vm/modules/operator/opimpl_f64.py
@@ -1,6 +1,7 @@
 from typing import TYPE_CHECKING, Any
 from spy.vm.b import B
-from spy.vm.object import W_Object, W_Type, W_F64, W_Bool
+from spy.vm.object import W_Object, W_Type
+from spy.vm.primitive import W_F64, W_Bool
 from . import OP
 if TYPE_CHECKING:
     from spy.vm.vm import SPyVM
