+++ spy/vm/modules/operator/opimpl_str.py
@@ -1,7 +1,8 @@
 from typing import TYPE_CHECKING
 from spy.vm.b import B
+from spy.vm.primitive import W_I32
 from spy.vm.str import W_Str
-from spy.vm.object import W_I32, W_Bool
+from spy.vm.primitive import W_Bool
 from . import OP
 if TYPE_CHECKING:
     from spy.vm.vm import SPyVM
