+++ spy/vm/modules/rawbuffer.py
@@ -4,10 +4,10 @@ SPy `rawbuffer` module.
 
 from typing import TYPE_CHECKING
 import struct
-from spy.vm.primitive import W_Void
+from spy.vm.primitive import W_F64, W_I32, W_Void
 from spy.vm.b import B
 from spy.vm.object import spytype
-from spy.vm.w import W_Func, W_Type, W_Object, W_I32, W_F64, W_Str
+from spy.vm.w import W_Func, W_Type, W_Object, W_Str
 from spy.vm.registry import ModuleRegistry
 if TYPE_CHECKING:
     from spy.vm.vm import SPyVM
