+++ spy/vm/str.py
@@ -1,10 +1,11 @@
 from typing import TYPE_CHECKING, Any
 from spy.llwasm import LLWasmInstance
 from spy.fqn import QN
-from spy.vm.object import W_Object, W_Type, W_Dynamic, spytype, W_I32
+from spy.vm.object import W_Object, W_Type, W_Dynamic, spytype
 from spy.vm.builtin import builtin_func
 from spy.vm.opimpl import W_OpImpl, W_OpArg
 from spy.vm.list import W_List
+from spy.vm.primitive import W_I32
 if TYPE_CHECKING:
     from spy.vm.vm import SPyVM
